The 'easy' part of #40

+++ co/dataIStreamQueue.cpp
@@ -44,7 +44,7 @@ DataIStreamQueue::~DataIStreamQueue()
 
 ObjectDataIStream* DataIStreamQueue::tryPop()
 {
-    QueuedStream stream( 0, (ObjectDataIStream*)0 );
+    QueuedStream stream( uint128_t( 0 ), (ObjectDataIStream*)0 );
     _queued.tryPop( stream );
     return stream.second;
 }
